project/requirements.txt
Django>=5.0
openai
python-dotenv
djangorestframework
requests
gunicorn
